# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get all todos",
                "description": "Display a listing of the todos.",
                "operationId": "dd687d49d7f86d473ce56a7806776fc4",
                "responses": {
                    "200": {
                        "description": "List of todos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "description": "Store a newly created todo in storage.",
                "operationId": "6f5da079c2dace66df52625276205a32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo created successfully"
                    }
                }
            }
        },
        "/api/todos/{id}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get a specific todo by ID",
                "description": "Display the specified todo.",
                "operationId": "5ee0e89cdc77f538ae94769b0b5cfae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo details"
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update a specific todo by ID",
                "description": "Update the specified todo in storage.",
                "operationId": "0e5cec090fe772f5dcb769dd9e38717c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo updated successfully"
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a specific todo by ID",
                "description": "Remove the specified todo from storage.",
                "operationId": "2f51ac0f120b8a6c7a8092bbd18174f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Todo deleted successfully"
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            }
        }
    }
}